requirements.txt
streamlit
pandas
numpy
scipy
plotly
requests
google-generativeai